criando componentes novos e organizando um pouco melhor o código
