+++ src/app/links/page.js
@@ -1,0 +1,15 @@
+import Header from "@/components/Header";
+import Navbar from "@/components/Navbar";
+
+export default function Page(){
+
+    return(
+        <div>
+            <Header />
+
+            <main>
+                <Navbar />
+            </main>
+        </div>
+    );
+}
