+++ styles/homePage.css
@@ -4,28 +4,6 @@ div {
     height: 100dvh;
 }
 
-.title {
-    text-align: center;
-    /* border: 1px solid #000; */
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    font-style: normal;
-    font-size: 25px;
-}
-
-p {
-    padding: 0 10px;
-}
-
-header {
-    display: flex;
-    color: #FFF;
-    background-color: #564D88;
-    height: 13%;
-    padding-left: 30px;
-}
-
 main {
     background-color: #F3F0F0;
     height: 87%;
