+++ styles/links.css
@@ -1,0 +1,4 @@
+
+main {
+    background-color: #F3F0F0;
+}
