+++ styles/navbarComponent.css
@@ -1,0 +1,20 @@
+
+nav {
+    height: 20%;
+    background-color: #F3F0F0;
+    border: 1px solid #000;
+    display: flex;
+    align-items: center;
+}
+
+.itemsContainer {
+    display: flex;
+    justify-content: space-around;
+    align-items: center;
+    width: 100%;
+    border: 1px solid #000;
+}
+
+.itemsContainer li {
+    border: 1px solid #000;
+}
